Add avaje tests for email and pattern

+++ validator-constraints/src/main/java/io/avaje/validation/constraints/Max.java
@@ -10,7 +10,7 @@ public @interface Max {
 
     Class<?>[] groups() default {};
 
-    double value();
+    long value();
 
     @Target({ElementType.METHOD, ElementType.FIELD})
     @Retention(RetentionPolicy.RUNTIME)
